+++ src/plot/layer/geom/spatial.rs
@@ -23,10 +23,6 @@ impl GeomTrait for Spatial {
         }
     }
 
-    fn needs_stat_transform(&self, _aesthetics: &Mappings) -> bool {
-        true
-    }
-
     fn apply_stat_transform(
         &self,
         query: &str,